requirements.txt
Flask
pymysql
python-dotenv
gunicorn
cryptography